Add possibility to hide notes
This commit enables the possibility to hide notes assets from the
environment configuration. The field `hidden` now indicates if the asset
should be invisible, which is not configurable for now.

We separate the assets from the usual environment.ts to separate the
assets only when it comes to testing. The logger has been made
independent from the environment and populates the level via
`isDevMode()`.

If everything works as intended can be verified by using the new target
`npm run start:prod`.

Signed-off-by: Sascha Grunert <sgrunert@suse.com>